requirements.txt
# CORE QA & TESTING
# -----------------
pytest>=7.0.0
pandas>=1.5.0
httpx>=0.24.0

# CODE QUALITY
# ------------
black>=23.0.0
flake8>=6.0.0
mypy>=1.8.0

# AI INTEGRATION
# -------------------------
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# BACKEND HTTP API
# ----------------
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# SECURITY & AUTHENTICATION
# -------------------------
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# INFRASTRUCTURE & PERSISTENCE
# ----------------------------
redis>=5.0.0
arq>=0.25.0
boto3>=1.34.0
tenacity>=8.2.0

# OBSERVABILITY & TELEMETRY
# -------------------------
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
# FIXED: Allow beta versions for FastAPI instrumentation
opentelemetry-instrumentation-fastapi>=0.40b0

# GRAMMATICAL FRAMEWORK
# ---------------------
# NOTE: This usually fails on Windows without SWIG/C-libs.
# We keep it for Docker, but for local Windows 'pip install', you might need to comment it out
# if you don't have the binary wheel.
pgf==1.1

# LEGACY / UTILS
# --------------
requests>=2.31.0